Param gov config changes (#585)
* Moved changes from old branch because rebase err

* Taking out unnecessary logic
