+++ pages/dao/[symbol]/params/index.tsx
@@ -4,12 +4,7 @@ import PreviousRouteBtn from '@components/PreviousRouteBtn'
 import useRealm from '@hooks/useRealm'
 import { fmtMintAmount } from '@tools/sdk/units'
 import { capitalize } from '@utils/helpers'
-import { getTreasuryAccountItemInfoV2 } from '@utils/treasuryTools'
 import useGovernanceAssetsStore from 'stores/useGovernanceAssetsStore'
-import {
-  getFormattedStringFromDays,
-  SECS_PER_DAY,
-} from 'VoteStakeRegistry/tools/dateTools'
 import Tabs from '@components/Tabs'
 import Select from '@components/inputs/Select'
 import Button from '@components/Button'
@@ -17,16 +12,17 @@ import useGovernanceAssets from '@hooks/useGovernanceAssets'
 
 import RealmConfigModal from './RealmConfigModal'
 import GovernanceConfigModal from './GovernanceConfigModal'
-import { VoteTipping } from '@solana/spl-governance'
 import { tryParsePublicKey } from '@tools/core/pubkey'
 import { getAccountName } from '@components/instructions/tools'
 import useWalletStore from 'stores/useWalletStore'
 import SetRealmAuthorityModal from './SetRealmAuthorityModal'
-import BigNumber from 'bignumber.js'
-import { AccountType } from '@utils/uiTypes/assets'
+
+import ParamsView from './components/ParamsView'
+import AccountsView from './components/AccountsView'
+import StatsView from './components/StatsView'
 
 const Params = () => {
-  const { realm, mint, councilMint, ownVoterWeight } = useRealm()
+  const { realm, mint } = useRealm()
   const wallet = useWalletStore((s) => s.current)
   const { canUseAuthorityInstruction, assetAccounts } = useGovernanceAssets()
   const governancesArray = useGovernanceAssetsStore((s) => s.governancesArray)
@@ -83,6 +79,7 @@ const Params = () => {
       setActiveGovernance(governancesArray[0])
     }
   }, [governancesArray])
+
   return (
     <div className="grid grid-cols-12 gap-4">
       {isRealmProposalModalOpen && (
@@ -118,30 +115,30 @@ const Params = () => {
             <>
               <div className="border border-fgd-4 col-span-1 p-4 rounded-md">
                 <h2>Addresses</h2>
-                <DisplayField
+                <AddressField
                   padding
                   label="Pubkey"
                   val={realm?.pubkey.toBase58()}
                 />
-                <DisplayField
+                <AddressField
                   padding
                   label="Authority"
                   val={realmAccount?.authority?.toBase58()}
                 />
-                <DisplayField
+                <AddressField
                   padding
                   label="Owner"
                   val={realm?.owner.toBase58()}
                 />
                 {communityMint && (
-                  <DisplayField
+                  <AddressField
                     padding
                     label="Community Mint"
                     val={communityMint}
                   />
                 )}
                 {councilMintPk && (
-                  <DisplayField
+                  <AddressField
                     padding
                     label="Council Mint"
                     val={councilMintPk}
@@ -162,13 +159,13 @@ const Params = () => {
               <div className="border border-fgd-4 col-span-1 p-4 rounded-md">
                 <h2 className="flex items-center">Config </h2>
                 {communityMintMaxVoteWeightSource && (
-                  <DisplayField
+                  <AddressField
                     padding
                     label="Community mint max vote weight source"
                     val={`${communityMintMaxVoteWeightSource.fmtSupplyFractionPercentage()}%`}
                   />
                 )}
-                <DisplayField
+                <AddressField
                   padding
                   label="Min community tokens to create governance"
                   val={
@@ -179,14 +176,14 @@ const Params = () => {
                     )
                   }
                 />
-                <DisplayField
+                <AddressField
                   padding
                   label="Use community voter weight add-in"
                   val={getYesNoString(
                     realmConfig?.useCommunityVoterWeightAddin
                   )}
                 />
-                <DisplayField
+                <AddressField
                   padding
                   label="Use max community voter weight add-in"
                   val={getYesNoString(
@@ -267,262 +264,23 @@ const Params = () => {
                   <Tabs
                     activeTab={activeTab}
                     onChange={(t) => setActiveTab(t)}
-                    tabs={['Params', 'Accounts']}
+                    tabs={['Params', 'Accounts', 'Statistics']}
                   />
                 ) : null}
-                {activeTab === 'Params' ? (
-                  <>
-                    <DisplayField
-                      label="owner"
-                      padding
-                      val={activeGovernance.owner.toBase58()}
-                    />
-                    {realmAccount?.authority?.toBase58() ===
-                      activeGovernance.pubkey.toBase58() && (
-                      <DisplayField
-                        label="Realm Authority"
-                        padding
-                        val={'Yes'}
-                      />
-                    )}
-                    <DisplayField
-                      label="Proposals Count"
-                      padding
-                      val={activeGovernance.account.proposalCount}
-                    />
-                    <DisplayField
-                      label="Voting Proposals Count"
-                      padding
-                      val={activeGovernance.account.votingProposalCount}
-                    />
-                    <DisplayField
-                      label="Max Voting Time"
-                      padding
-                      val={getFormattedStringFromDays(
-                        activeGovernance.account.config.maxVotingTime /
-                          SECS_PER_DAY
-                      )}
-                    />
-                    {communityMint && (
-                      <DisplayField
-                        label="Min community tokens to create a proposal"
-                        padding
-                        val={fmtMintAmount(
-                          mint,
-                          activeGovernance.account.config
-                            .minCommunityTokensToCreateProposal
-                        )}
-                      />
-                    )}
-                    {councilMint && (
-                      <DisplayField
-                        label="Min council tokens to create a proposal"
-                        padding
-                        val={fmtMintAmount(
-                          councilMint,
-                          activeGovernance.account.config
-                            .minCouncilTokensToCreateProposal
-                        )}
-                      />
-                    )}
-                    <DisplayField
-                      label="Min Instruction Holdup Time"
-                      padding
-                      val={`${activeGovernance.account.config.minInstructionHoldUpTime} secs`}
-                    />
-                    <DisplayField
-                      label="Proposal Cool-off Time"
-                      padding
-                      val={`${activeGovernance.account.config.proposalCoolOffTime} secs`}
-                    />
-                    <DisplayField
-                      label="Vote Threshold Percentage"
-                      padding
-                      val={`${activeGovernance.account.config.voteThresholdPercentage.value}%`}
-                    />
-                    <DisplayField
-                      label="Vote Tipping"
-                      padding
-                      val={
-                        VoteTipping[
-                          activeGovernance.account.config.voteTipping as any
-                        ]
-                      }
-                    />
-                    <div className="flex">
-                      <Button
-                        disabled={
-                          !ownVoterWeight.canCreateProposal(
-                            activeGovernance.account.config
-                          )
-                        }
-                        tooltipMessage={
-                          !ownVoterWeight.canCreateProposal(
-                            activeGovernance.account.config
-                          )
-                            ? 'Please connect wallet with enough voting power to create governance config proposals'
-                            : ''
-                        }
-                        onClick={openGovernanceProposalModal}
-                        className="ml-auto"
-                      >
-                        Change config
-                      </Button>
-                    </div>
-                  </>
-                ) : (
-                  <div className="space-y-3">
-                    {assetAccounts
-                      .filter(
-                        (x) =>
-                          x.governance.pubkey.toBase58() ===
-                          activeGovernance.pubkey.toBase58()
-                      )
-                      .map((x) => {
-                        const info = getTreasuryAccountItemInfoV2(x)
-                        if (x.isToken || x.isSol) {
-                          return (
-                            <div
-                              className="bg-bkg-1 p-4 pb-2 rounded-md"
-                              key={x.pubkey.toBase58()}
-                            >
-                              <DisplayField
-                                bg={false}
-                                label="Name"
-                                val={info.name}
-                              />
-                              <DisplayField
-                                bg={false}
-                                label="Address"
-                                val={x.extensions?.transferAddress?.toBase58()}
-                              />
-                              <DisplayField
-                                bg={false}
-                                label="Balance"
-                                val={
-                                  <div className="flex items-center">
-                                    {info.logo && (
-                                      <img
-                                        className="h-4 mr-1 w-4"
-                                        src={info.logo}
-                                      />
-                                    )}
-                                    <span>{`${info.amountFormatted} ${
-                                      info.info?.symbol || ''
-                                    }`}</span>
-                                  </div>
-                                }
-                              />
-                              <DisplayField
-                                bg={false}
-                                label="Type"
-                                val={AccountType[x.type]}
-                              />
-                              {x.type !== AccountType.SOL && (
-                                <DisplayField
-                                  label="Mint"
-                                  bg={false}
-                                  val={x.extensions.mint?.publicKey.toBase58()}
-                                />
-                              )}
-                            </div>
-                          )
-                        }
-
-                        if (x.type === AccountType.NFT) {
-                          return (
-                            <div
-                              className="bg-bkg-1 p-4 pb-2 rounded-md"
-                              key={x.pubkey.toBase58()}
-                            >
-                              <DisplayField
-                                bg={false}
-                                label="Type"
-                                val={AccountType[x.type]}
-                              />
-                              <DisplayField
-                                bg={false}
-                                label="Address"
-                                val={x.extensions?.transferAddress?.toBase58()}
-                              />
-                            </div>
-                          )
-                        }
-                        if (x.type === AccountType.MINT) {
-                          return (
-                            <div
-                              className="bg-bkg-1 p-4 pb-2 rounded-md"
-                              key={x.pubkey.toBase58()}
-                            >
-                              <DisplayField
-                                bg={false}
-                                label="Type"
-                                val={AccountType[x.type]}
-                              />
-                              <DisplayField
-                                bg={false}
-                                label="Pubkey"
-                                val={x.extensions.mint?.publicKey.toBase58()}
-                              />
-                              <DisplayField
-                                bg={false}
-                                label="Decimals"
-                                val={x.extensions.mint?.account.decimals}
-                              />
-                              <DisplayField
-                                bg={false}
-                                label="Mint Authority"
-                                val={x.extensions.mint?.account.mintAuthority?.toBase58()}
-                              />
-                              <DisplayField
-                                bg={false}
-                                label="Supply"
-                                val={new BigNumber(
-                                  x.extensions.mint!.account.supply.toString()
-                                )
-                                  .shiftedBy(
-                                    -x.extensions.mint!.account.decimals || 0
-                                  )
-                                  .toFormat()}
-                              />
-                              <DisplayField
-                                bg={false}
-                                label="Is Initialized"
-                                val={getYesNoString(
-                                  x.extensions.mint?.account.isInitialized
-                                )}
-                              />
-                              {x.extensions.mint?.account.freezeAuthority ? (
-                                <DisplayField
-                                  bg={false}
-                                  label="Freeze Authority"
-                                  val={x.extensions.mint?.account.freezeAuthority?.toBase58()}
-                                />
-                              ) : null}
-                            </div>
-                          )
-                        }
-                        if (x.type === AccountType.PROGRAM) {
-                          return (
-                            <div
-                              className="bg-bkg-1 p-4 pb-2 rounded-md"
-                              key={x.pubkey.toBase58()}
-                            >
-                              <DisplayField
-                                bg={false}
-                                label="Type"
-                                val={AccountType[x.type]}
-                              />
-                              <DisplayField
-                                bg={false}
-                                label="Pubkey"
-                                val={x.pubkey.toBase58()}
-                              />
-                            </div>
-                          )
-                        }
-                      })}
-                  </div>
+                {activeTab === 'Params' && (
+                  <ParamsView
+                    activeGovernance={activeGovernance}
+                    openGovernanceProposalModal={openGovernanceProposalModal}
+                  />
+                )}
+                {activeTab === 'Accounts' && (
+                  <AccountsView
+                    activeGovernance={activeGovernance}
+                    getYesNoString={getYesNoString}
+                  />
+                )}
+                {activeTab === 'Statistics' && (
+                  <StatsView activeGovernance={activeGovernance} />
                 )}
               </div>
             ) : null}
@@ -535,7 +293,7 @@ const Params = () => {
   )
 }
 
-const DisplayField = ({ label, val, padding = false, bg = false }) => {
+export const AddressField = ({ label, val, padding = false, bg = false }) => {
   const pubkey = isNaN(val) && tryParsePublicKey(val)
   const name = pubkey ? getAccountName(pubkey) : ''
   return (
@@ -554,6 +312,26 @@ const DisplayField = ({ label, val, padding = false, bg = false }) => {
         ) : (
           <div>{val}</div>
         )}
+      </div>
+    </div>
+  )
+}
+
+export const NumberField = ({
+  label,
+  val = 0,
+  padding = false,
+  bg = false,
+}) => {
+  return (
+    <div
+      className={`flex flex-col mb-2 ${bg ? 'bg-bkg-1' : ''} ${
+        padding ? 'py-1' : ''
+      }`}
+    >
+      <div className="text-xs text-fgd-3">{capitalize(label)}</div>
+      <div className="text-sm break-all">
+        <div>{val}</div>
       </div>
     </div>
   )
